session-3/requirements.txt
pandas
numpy
scikit-learn
mlflow
joblib
matplotlib
seaborn
fastapi
uvicorn
requests
xgboost
lightgbm
shap
streamlit
scipy
pillow
